chore: defensive hardening improvements
- EnchantingStatRegistry: Added volatile idempotency guard to bootstrap.
- EnchantingRecipeRegistry: Added volatile to registered flag.
- EnchantingStats: Added validation to codec for non-negative values.
- EnchantmentLibraryBlockEntity: Improved precision in maxLevelAffordable using bitwise logic.
- AnvilDispatcher: Switched to CopyOnWriteArrayList and removed manual synchronization.
- ArmorTickHandler: Changed tickCounter to long to prevent overflow.
- MeridianEnchantmentMenu: Switched to type-safe lists for client clues.
- TooltipFormatter: Moved to neutral package com.rfizzle.meridian.util.
- EnchantingInfoScreen: Changed PowerSlider visibility to private.
- EnchantmentInfo: Fixed unnecessary double-cast in STREAM_CODEC.
- ScrapTomeHandler: Replaced java.util.Random with RandomSource.
- JeiEnchantingCategory: Updated MAX_LINES for better layout.

+++ src/main/java/com/rfizzle/meridian/event/ArmorTickHandler.java
@@ -26,7 +26,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public final class ArmorTickHandler {
 
-    private static int tickCounter = 0;
+    private static long tickCounter = 0;
 
     private static final Map<BlockPos, Long> cinderwalkBlocks = new ConcurrentHashMap<>();
     private static final int CINDERWALK_REVERT_TICKS = 80;
